img: new flip_x and flip_y parameters, fix extract=false
